Load class in EnumConverter using context classloader.
This allows loading correctly in runtimes that scope application code to application classloaders.
If JNoSQL classes are included in such a runtime, the class is loaded from the application instead of from the runtime's shared classloader.
It still works correctly for flat classpath because the context classloader is set to the bootstrap classloader by default.